Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2392 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/ui/framewindow.cpp
@@ -566,7 +566,8 @@ LRESULT qs::FrameWindow::onInitMenuPopup(HMENU hmenu,
 				break;
 			
 			if (mii.dwItemData != 0) {
-				DynamicMenuCreator* pMenuCreator = getDynamicMenuCreator(mii.dwItemData);
+				DynamicMenuCreator* pMenuCreator = getDynamicMenuCreator(
+					static_cast<DWORD>(mii.dwItemData));
 				if (pMenuCreator)
 					nIndex = pMenuCreator->createMenu(hmenu, nIndex);
 				else
